BLE Host - Read BR/EDR buffer size at startup.
If the LE buffer size is 0, the controller shares the same buffers for
LE and BR/EDR.  In this case, use the BR/EDR values.

+++ net/nimble/host/src/ble_hs_conn_priv.h
@@ -63,7 +63,7 @@ struct ble_hs_conn {
      * Count of packets sent over this connection that the controller has not
      * transmitted or flushed yet.
      */
-    uint8_t bhc_outstanding_pkts;
+    uint16_t bhc_outstanding_pkts;
 
     /** Queue of outgoing packets that could not be sent. */
     STAILQ_HEAD(, os_mbuf_pkthdr) bhc_tx_q;
