+++ net/nimble/host/src/ble_hs_hci_priv.h
@@ -69,7 +69,7 @@ struct ble_hs_hci_ext_conn_params {
 };
 #endif
 
-extern uint8_t ble_hs_hci_avail_pkts;
+extern uint16_t ble_hs_hci_avail_pkts;
 
 int ble_hs_hci_cmd_tx(uint16_t opcode, void *cmd, uint8_t cmd_len,
                       void *evt_buf, uint8_t evt_buf_len,
@@ -161,7 +161,7 @@ int ble_hs_hci_cmd_le_conn_param_neg_reply(
 void ble_hs_hci_cmd_build_le_start_encrypt(const struct hci_start_encrypt *cmd,
                                            uint8_t *dst, int dst_len);
 int ble_hs_hci_set_buf_sz(uint16_t pktlen, uint16_t max_pkts);
-void ble_hs_hci_add_avail_pkts(uint8_t delta);
+void ble_hs_hci_add_avail_pkts(uint16_t delta);
 
 uint16_t ble_hs_hci_util_handle_pb_bc_join(uint16_t handle, uint8_t pb,
                                            uint8_t bc);
